Refactor data indexing structure.
Create a service provider to register the default interfaces (RepositoryReader and Repository)
So the IoC loads the appropiate one.

This frees the start/global.php file from some unneeded code.

There's a negligible performance boost, as the indexers are no longer
instantiated on each request, but only when rebuilding the database.