There is no need to set an expiration on searches.
As the database only changes when it's rebuilt, and the cache is cleared
when it's rebuild.

Checking this I noticed there's an attack vector, doing requests with
a permutation of different search values, they will all be stored until
the cache runs out of space. This is not really introduced in this commit,
but it's made more visible.

I need to write a mechanism to prevent this, meanwhile...
please do not attack the search.