Searches were being stored in the same cache as the game database cache.
When somebody performed a search, the file would be owned by the web server,
while the process to rebuild the database might be with a different user.

This caused problems when rebuilding the database in the devel copy of the
app, as the user cannot delete the search results, resulting in an
inconsistent internal database.

using \Cache it means it will be stored in the default cache, which
currently is memcache.